+++ src/components/scroll-view/scroll-view.module.css.d.ts
@@ -6,3 +6,7 @@
  * actual or intended publication of such source code.
  */
 export const root: string;
+
+export const inverted: string;
+
+export const scrollbarsAlwaysShown: string;