+++ src/components/scroll-view/scroll-view.tsx
@@ -6,27 +6,46 @@
  * actual or intended publication of such source code.
  */
 import classNames from 'classnames';
-import scrollbarSize from 'dom-helpers/util/scrollbarSize';
 import * as React from 'react';
 import SimpleBar from 'simplebar-react';
 import 'simplebar/dist/simplebar.min.css';
 
+import {DarkContext} from '../../utils/dark-context';
+
 import * as styles from './scroll-view.module.css';
 
 /** Props of ScrollView component */
-export type Props = React.ComponentPropsWithoutRef<'div'>;
+export type Props = React.ComponentPropsWithoutRef<'div'> & {
+    /**
+     * Defines when and how to show scrollbars. Possible values:
+     *
+     * `auto` — scrollbars show up on hover or drag
+     *
+     * `always` — scrollbars are always shown
+     */
+    showScrollbars?: 'auto' | 'always';
+};
+
+const simplebarOptions = {
+    autoHide: false
+};
 
 /**
  * A wrapper component over [SimpleBar](https://github.com/Grsmto/simplebar/tree/master/packages/simplebar-react) that provides crossplatform experience for scrollable views.
- *
- * Note: native scrollbars will be used if they're displayed in "overlay" mode.
  */
-export const ScrollView = (props: Props) => {
-    const scrollbarsShown = Boolean(scrollbarSize());
+export const ScrollView = ({className, showScrollbars = 'auto', ...rest}: Props) => {
+    const inverted = React.useContext(DarkContext);
 
-    return scrollbarsShown ? (
-        <SimpleBar {...props} className={classNames(styles.root, props.className)} />
-    ) : (
-        <div {...props} />
+    return (
+        <SimpleBar
+            {...rest}
+            className={classNames(
+                styles.root,
+                {[styles.inverted]: inverted},
+                {[styles.scrollbarsAlwaysShown]: showScrollbars === 'always'},
+                className
+            )}
+            {...simplebarOptions}
+        />
     );
 };